requirements.txt
flask
kafka-python
pyspark
psycopg2
flask
sqlalchemy
python-dotenv